Update learning rate schedule in training/default.yaml
This updates the default values for the learning rate schedule. Specifically, the local learning rate is increased to 8x the original learning rate. In addition, the number of iterations is reduced from 300k to 150k. This has led to similar accuracy in stretched grid experiments using the graph transformer architecture, while requiring only 50% of the GPU compute.